+++ router/throttler/internal/pickup/adaptive/adaptive_all_event_types.go
@@ -25,29 +25,20 @@ func NewAllEventTypesThrottler(destType, destinationID string, algorithm Algorit
 			fmt.Sprintf(`Router.throttler.%s.%s.minLimit`, destType, destinationID),
 			fmt.Sprintf(`Router.throttler.%s.minLimit`, destType),
 			`Router.throttler.minLimit`,
-			// TODO: delete the following deprecated keys in the future
-			fmt.Sprintf(`Router.throttler.adaptive.%s.%s.minLimit`, destType, destinationID),
-			fmt.Sprintf(`Router.throttler.adaptive.%s.minLimit`, destType),
-			`Router.throttler.adaptive.minLimit`,
 		),
 		maxLimit: maxLimitFunc(c, destType, destinationID,
 			[]string{
 				fmt.Sprintf(`Router.throttler.%s.%s.maxLimit`, destType, destinationID),
 				fmt.Sprintf(`Router.throttler.%s.maxLimit`, destType),
 				`Router.throttler.maxLimit`,
-				// TODO: delete the following deprecated keys in the future
-				fmt.Sprintf(`Router.throttler.adaptive.%s.%s.maxLimit`, destType, destinationID),
-				fmt.Sprintf(`Router.throttler.adaptive.%s.maxLimit`, destType),
-				`Router.throttler.adaptive.maxLimit`,
 			},
 		),
 		staticCost: c.GetReloadableBoolVar(true,
 			`Router.throttler.adaptiveIgnoreThrottlingCosts`,
-			`Router.throttler.adaptive.ignoreThrottlingCosts`, // TODO: delete this deprecated key in the future
 			`Router.throttler.ignoreThrottlingCosts`,
 		),
 
-		everyGauge: kitsync.NewOnceEvery(time.Second),
+		everyStats: kitsync.NewOnceEvery(200 * time.Millisecond),
 		limitFactorGauge: stat.NewTaggedStat("adaptive_throttler_limit_factor", stats.GaugeType, stats.Tags{
 			"destinationId": destinationID,
 			"destType":      destType,