+++ router/throttler/internal/pickup/switcher/switcher.go
@@ -2,6 +2,7 @@ package switcher
 
 import (
 	"context"
+	"time"
 
 	"github.com/rudderlabs/rudder-go-kit/config"
 	"github.com/rudderlabs/rudder-server/router/throttler/internal/types"
@@ -42,9 +43,19 @@ func (t *throttlerSwitcher) Shutdown() {
 	t.alternative.Shutdown()
 }
 
-// getLimit returns the limit of the currently active throttler.
-func (t *throttlerSwitcher) GetLimit() int64 {
-	return t.throttler().GetLimit()
+// GetLimitPerSecond returns the limit of the currently active throttler.
+func (t *throttlerSwitcher) GetLimitPerSecond() int64 {
+	return t.throttler().GetLimitPerSecond()
+}
+
+// GetEventType returns the event type of the currently active throttler.
+func (t *throttlerSwitcher) GetEventType() string {
+	return t.throttler().GetEventType()
+}
+
+// GetLastUsed returns the last used time of the currently active throttler.
+func (t *throttlerSwitcher) GetLastUsed() time.Time {
+	return t.throttler().GetLastUsed()
 }
 
 // throttler returns the currently active throttler based on the useAlternative config.